Bug#36562979: MySQL crash at VarintParse64()
The hash join buffer could in some cases have too little space for a
row to be written, so that data could be written outside of the
allocated buffer and cause erroneous behaviour.

This happened if the hash join buffer stored row ids from a table in
the MEMORY engine. The reason was that the maximum size per row was
calculated before the row id size (that is, handler::ref_length) had
been initialized in the MEMORY engine handler, so the default
ref_length for the handler class was used instead, which is
sizeof(my_off_t). This is usually half the correct size, which is
sizeof(HP_HEAP_POSITION).

Fixed by initializing ref_length already when the MEMORY engine
handler is constructed, so that it always has the correct value. Also
add more assertions to detect the problem earlier.

Change-Id: I6a7a675f2117b78c5b279bbba1abd8da641fad6e
(cherry picked from commit 05b23e971e0e779406043c15cc8b4b417a246c91)