Bug#36764973 MTR TEST innodb.parallel_read_kill FAILING IN MYSQL 8.0.37
Description
-----------
The test has been failing 8.0.37 onwards.

Background
-----------
Test creates a table with a secondary index and inserts 150K un-commited
records in it. Earlier Innodb was ignoring the optimizer hints but using
the clustered index to count the records in the table. This behavior was
reverted through Bug#35952353 [SELECT COUNT(*) degraded performance on
8.0 compared to 5.7] in 8.0.37. Now the query execution is quicker than
before since it is using secondary index to count the records. Test
doesn't have any dependency on the secondary index, therefore, removing
the secondary index key from the table would restore the intent of the
test.

Fix
----
* Removed the secondary index key from the table.

* Prepare the table of 150K records through 8 user connections, each
  of them inserts 18750 records. This reduces test execution duration
  to qualify it as a regular test than big test.

Change-Id: I14bd90cf765ea3ad200e2d83c962f648f8366111