Bug#36805335 ODR violations around mysys and vio
Build with ASAN and gcc reveals ODR violations in router:

- vio (key_memory_vio_ssl_fd)
- charsets (init_charsets)

which is caused by static libraries being linked into multiple shared
objects:

- mysys into harness, router_mysql, metadata_cache
- vio into harness and mysql_mysql

The linking against vio, mysys, ... is done to get access to
my_print_stacktrace and similar functions from mysys,
but that collides with mysys that's included in the mysqlclient library.

As mysqlclient is a static library too, the router ends up with _two_
insteances of mysys -> ODR violation.

Change
======

- build a router_mysqlclient shared library from the libmysql-parts
  that export all the symbols router needs.
- replace linking against the static mysqlclient, vio, strings and mysys
  with router_mysqlclient
- use get_charset() instead of direct address of my_charset_* to avoid
  garbage addresses on windows.

Change-Id: Ic413a73b1d5ddd41fae44bddf74e89b259907ca2