Bug#35343458  Do not attempt to drop FTS tables when parent tablespace is discarded
Given a table with an FTS index, when its tablespace is discarded, the
corresponding FTS tables are dropped. When performing an operation that
would clear out these tables, the code assumes that if the FTS index
exists in the table metadata, the FTS tables must also exist. The fix is
to check for this condition and skip emptying the tables.

Change-Id: I13844c92434bb0e4ad38ff13d6132d1ac5770d67