Bug #36808732 - Crash in restart after creating 10,000 tables
SYMPTOM:
- There is a crash that is occurring during server startup if there
are more than 8000 tablespaces

ROOT CAUSE:
- This was a regression introduced by the fix for Bug #33398681
- The function Validate_files::check() launches multiple scan threads
to validate the tablespaces if the number of tablespaces is greater
than FIL_SCAN_MAX_TABLESPACES_PER_THREAD at start up.
- Scan threads are normal threads and not registered to mysqld server
- One of the threads scanning the tablespace tried to access the
dd_client object, but it doesn't have the THD object associated with
it, therefore it crashes.

FIX:
- We removed the logic which needed dd_client object to a different
function dd_update_table_and_partitions_after_dir_change() which is
called by the regular mysql recovery thread.
- Bug #33398681 had introduced two more regressions in failing test
cases which are fixed in this commit:

(i) partition_innodb_tablespace.test
External data directory flag was not set for a partition in the
partitioned table because of which it was displaying the data
directory incorrectly in the "SHOW CREATE TABLE" command. Fixed this
by properly setting the flag for the partitioned tables which have
an external directory.

(ii) portability_tablespace_windows.test
Partition names were not converted to the windows format if the
paritioned table had a external data directory, because of which
they could not be recognized after restart. This has been fixed.

Change-Id: I6368bf9dd983cfd08c1479a1674a520ea5c47627