Bug#36758378: Signal 11 in Table_ref::is_external
Post-push fix.

The original fix set the LEX::has_external_tables() property only when
the query was offloaded to a secondary engine, but it is needed also
in the cases where a query accessing some external tables is executed
in the primary engine.

Fixed by moving the initialization of the property so that it's set
for both primary engine execution and secondary engine execution.

Change-Id: Ia2e7ac41b8befa75d6e57074428c2e4fe5e9e845