Bug#35553557 Table creation with foreign key assertion error
Problem:
When executing a CREATE TABLE ... AS SELECT... statement, an assertion
is encountered when the CREATE TABLE clause creates a table with a
foreign key relationship to a table referenced in the SELECT clause.

Root cause:
When foreign keys are involved in a DML SQL statement, the
Foreign_key_parents_invalidator::invalidate method is used to ensure
that tables in its parent_map have their data dictionary cache entries
invalidated. The original technique used for this invalidation was to
close such tables.  Unfortunately, in the particular statement that
triggers this bug, a table in the parent_map was also being closed by
the Query_expression::cleanup. As the table metadata had already been
"trashed" by the first table close operation, an attempt to access this
"trashed" memory by the second table close operation resulted in an
assertion failure (in a debug build) or a memory violation in a released
build.

Solution:
Fix by skipping table calls to mysql_ha_flush_table and
close_all_tables_for_name in Foreign_key_parents_invalidator::invalidate
when the table name to be closed can be found in
Query_result_create::select_tables. This operation is implemented by
extending the Foreign_key_parents_invalidator::m_parent_map to hold a
flag indicating whether a table should be invalidated and closed or
invalidated and marked for reopen. The value of this flag is defaulted
to invalidate and close when tables are added to the map, but it can be
overriden both when adding tables to the map and subsequently by invoking
a new Foreign_key_parents_invalidator function mark_for_reopen_if_added.
This new function searches for the given table name in the map and if
found sets the corresponding flag to invalidate and reopen.

The Foreign_key_parents_invalidator::invalidate function now inspects
the new flag in the m_parent_map and uses the table data cache flag
TDC_RT_MARK_FOR_REOPEN for those tables marked to be invalidated and
reopened instead of closing the table. Thus when the
Query_expression::cleanup function closes tables it no longer encounters
"trashed" memory values, and the previously "double deleted" table is
instead handled correctly - the TDC_RT_MARK_FOR_REOPEN that has been
applied to it ensures that its metadata is up to date when it is next
opened.

Change-Id: Ieb8dc95f067e93b7ae2254cfbfb9507576e17573