Bug#36635930 - Router crashes on mutex::try_lock() in http_server.so
Description
===========
The `http_server` plugin crashes with following callstack:
13  libc++.1.dylib                      0x000000019a52e72c std::__1::mutex::try_lock() + 0
14  http_server.so                      0x0000000105a10870 std::__1::unique_lock<....
15  http_server.so                      0x0000000105a107ac std::__1::unique_lock<....
16  http_server.so                      0x0000000105a2cc34 http::base::Connection<
                         net::tls::TlsStream<net::basic_stream_socket<net::ip::tcp>>::
                         on_net_send(std::__1::error_code const&, unsigned long) + 60

The behavior of http-plugin was changed recently to read and write simultaneously,
and the code that handless IO callback is not ready for that, any io-error tries
to close the connection (deletes the object) but the reference to the object
is keept in second IO callback.

Fix
===
Deallocate objects when both IO callbacks (read/write) ended.

Change-Id: I150b5e2c8901466679a7a898309457226b3cb9ea