Bug#36825088 get_charset_by_name("utf16le_bin") rather than global symbols
Fix a few places where we do
   extern CHARSET_INFO my_charset_utf16le_bin;
rather than
   get_charset_by_name("utf16le_bin", MYF(0))
which is the proper way to use character sets/collations.

Also change name of argument in get_charset_by_name to
'collation_name' to make it clearer that the expected actual argument
is a collation name rather than a character set name.

Change-Id: I687b0c1726873c95ec89bc47538335db8771762e