Bug#35497623: SHOW INDEXES with generated column may access wrong key-part
In SHOW INDEXES we iterate over the KEY_PART_INFOs
of a KEY, and if it's a functional index, we try
to show its gcol_info (as in, "generated column info").

Unfortunately in the unpatched server, we try to obtain
that gcol_info from the first key-part in the key,
rather than from the current key-part.

This patch rectifies this.

While we're at it, we also make Column_name NULLable,
seeing as we occasionally set it to NULL (as we do in
the test case added by this patch). This aligns the
result of SHOW INDEXES for temporary tables with that
of regular tables.

Change-Id: I021c905150e38cd84739e5f4f69494781fb3d88e