Bug#36829819 MTR test mfa_bugs_debug hangs on modern Ubuntu/Fedora platforms
The perl utility wait_for_pattern_in_file.inc does not work for newer
versions of Perl.  Once we see EOF, we stop reading input, and the
.mtr test eventually times out.

The fix is to reset the EOF marker for the input file.

Excerpt from perl documentation:

https://perldoc.perl.org/functions/seek
seek FILEHANDLE,POSITION,WHENCE

A WHENCE of 1 (SEEK_CUR) is useful for not moving the file position:

seek($fh, 0, 1)

This is also useful for applications emulating tail -f. Once you hit
EOF on your read and then sleep for a while, you (probably) have to
stick in a dummy seek to reset things. The seek doesn't change the
position, but it does clear the end-of-file condition on the handle,
so that the next readline FILE makes Perl try again to read
something. (We hope.)

Change-Id: Ic93be7583715dd62d2b691530c065e8526685979
(cherry picked from commit ac9ad390a3907ddcf46d658f353b357d4482469d)