Bug#36715239: Assertion `found_a_plan || is_secondary_engine' failed. [1/2, fix, noclose]
The hypergraph optimizer does not find a plan for some queries that
contain a semijoin on the inner side of an outer join.

The problem is that ProposeHashJoin() refuses to create hash joins for
certain left joins with a semijoin on the right side.
AllowNestedLoopJoin() sees that the corresponding nested loop join is
not a good candidate (since it cannot use indexes for the join
condition), and it doesn't see that the corresponding hash join was
rejected, so it rejects the nested loop join too. Because of this, no
plan is found.

Fixed by factoring out the checks for whether to allow a hash join in
a separate function that can be shared between ProposeHashJoin() and
AllowNestedLoopJoin(), so that they always agree on whether a hash
join is possible.

Change-Id: I81ef15b2eac52db5ba029891fa281251df7f228e