BUG#36820139 - Missing primary key field in MRS JSON response when inserting a row
Description
===========
User should be able to insert a row into a table available as a database object
in a MySQL REST Service instance, using an HTTP POST request,
where the body contains a JSON object that includes the fields and corresponding
values required for each table column to create a new row.

For example, to insert a new row in the "actor" table in the Sakila sample database
available as a database object in a "myService" REST service, a valid request could
contain the following JSON object in the body:

  {
    "firstName": "FOO",
    "lastName": "BAR"
  }

The database object contains two additional fields ("actorId", which is the primary key,
and "lastUpdated") that are auto-generated, these fields are not required to be part of
the JSON object. However, I expect the response to include a JSON object that contains
these fields and the values that have been assigned to them by omission.

Fix
===
The MRS plugin while fetching data from metadata-schema, was fetching `allow_sorting`
as `no_check`, which was leading to wrong behavior while generating response documents.

Change-Id: Id4c60d9f271a832ea994280658de531245ae3e00