Bug#36775485: Hypergraph optimizer does not set SUM_ROWS_EXAMINED in ps.events_statements_summary_by_digest for index range scans, and some index lookup [2/2, fix]
Fix the bug by making the hypergraph optimizer mark all base table
access paths with the "count_examined_rows" flag. This makes
sum_rows_examined get updated for all reads from base tables, also
when using index range scans and index lookups, which were missing
before.

It is fixed by updating the flag in one single place, in
CostingReceiver::ApplyPredicatesForBaseTable() which is called on
every base table access path.

Code that sets the flag other places in the hypergraph optimizer is
removed, as it is no longer necessary. Code that clears the flag is
also removed, as the AccessPath constructor already initializes the
flag with the value false.

Test cases that were previously disabled because of this bug, are
enabled.

Change-Id: I5600308f7fb60d2fcfda3cbd7d7325a2d1faf65a