Avoid latency glitches
When sending to another node and not being able to send all the data
it is important to make sure it gets sent anyways. Currently this
relies on the wakeup timer in the send thread. However this might
delay the next send on a ready transporter by as much as 1 millisecond
and possibly even 10 millisecond dependent on how the OS handles
timer interrupts.

A better approach is to immediately wakeup the responsible send
thread and have this thread take care of the sending and thus
ensuring it is sent as soon as is possible with the send thread.