Bug#36715239: Assertion `found_a_plan || is_secondary_engine' failed. [2/2, use hash join]
The previous patch made the hypergraph optimizer come up with a plan
for the query that could not be planned. However, the plan it came up
with was not ideal, as it used a nested loop left join whose inner
table could not use an index to restrict the number of rows read.

The reason for this was a restriction on using hash joins on certain
queries having a semijoin inside an outer join. The restriction was
added in bug#34079278. The intention was to avoid some problematic
hash joins for execution in secondary engines. The issue didn't affect
primary engines, so the check in AllowHashJoin() can be made a little
looser and allow these hash joins when executing in a primary engine.

Change-Id: I7dd172aa05a29125f3eefa8b8de86dcee5f87649