Bug#36845760: Allow unindexed nested loop join with LIMIT 1 on the outer table in hypergraph
The fix for bug#36625968 was a bit too strict and disallowed nested
loop join for some semijoins that were transformed to an inner join
with a LIMIT 1 on one of the tables. If the LIMIT 1 is on the left
side of a nested loop join, the nested loop join is guaranteed to read
each side of the join at most once, so it's just as safe as the
corresponding hash join, and probably a little faster.

Fixed by loosening the check in AllowNestedLoopJoin() so that it
allows nested loop joins with a LIMIT 1 on the left hand side, even if
there are no indexes that can be used on the right hand side. Also
move the call to AllowNestedLoopJoin() after the semijoin to inner
join transformation has been performed, so that the LIMIT 1 is
actually seen by it.

Change-Id: Ifae79fed424453ad578a37ac39ff22359302ca8a