Bug#36846616 mysql_harness::dim does not scale
Benchmarking showed that at 64 parallel threads calling log_debug()
results in lock-contention around mysql_harness::DIM.

The DIM is storage for global pointers for the Config, the Logger, ...

It initialized at startup and read a lot while the application runs.

But the access to all fields is protected with a single mutex: if thread
A wants to get the pointer to Config, thread B can't get the pointer to
the Logger. Even though that's not needed.

Each get_* method also has to use a exclusive mutex as it may lazy init
the "to-be-get" object which only happens once. It would be better to
make get purely read-only and use a "std::shared_lock".

Change
======

- removed lazy instantiation in DIM
- replaced global single mutex with a std::shared_mutex per object in
  DIM to allow parallel "get" operations.
- removed all factory-methods as lazy instantiation is removed.
- refactored the DIM code

Change-Id: I7c28e46884f7a50f219603e33559edcc4dac773f