Bug#36941628 logging functions take too many locks
Profiling shows that the logging functions log_debug(), log_info(),
... take too many locks and filtering out a log message is too costly.

Locks
=====

Each time a log_debug() is called, it checks the registry for the Logger
of the current domain and logs the message in that logger. Checking the
registry, requires a Lock.

The code:

  log_debug("%s", "abc");
  log_debug("%s", "def");

checks the registry twice if there is a Logger for the current
LogDomain.

Better:

  DomainLogger logger;  // check registry once
  logger.debug("abc");
  logger.debug("def");

Filtering Log Messages
======================

If the log_level is 'info' the code:

  std::string msg = some_costly_func();

  log_debug("%s", msg.c_str());

leads to 'msg' being created, and then later discarded.

Better, call the costly debug-msg function only of the message actually
gets logged:

  DomainLogger logger;
  logger.debug([](){ return some_costly_func(); });

Change
======

- added DomainLogger which checks the Registry once, if one-or-more
  messages are logged.
- allow the DomainLogger to create the log-message if the log-message
  is actually getting logged.
- adapted the mock-server code to use DomainLogger.

Change-Id: Icd7b74ca09bbeb7a1e9080956233fea13386351f