Bug#36723117: Crash using a secondary index after dropping a column with               ALGORITHM=INSTANT
Analysis:
The server may abort when executing DELETE or UPDATE statement
after a column is dropped using INSTANT algorithm.

The server aborts during DELETE or UPDATE statement when the end
range comparison is triggered. In the section of the code where
the end range optimisation is performed, the offset and field_no
passed is that of the secondary index but the index record is that
of the clustered index. Since the table has a row_version due
to the INSTANT DROP, the field position in the offset returned
is invalid. This causes the server to abort during the field
position validation.

Fix:
Pass the secondary index record instead of the clustered index
for cases where the end range comparison algorithm is triggered.

Change-Id: Icb1108e35b3dbff0d83fb530c96fd16ba8544f86