Bug#36773414: Explain format=tree does not display               information about semijoin strategy used
The traditional explain format always displays the
semijoin strategy used in the query. However, this
information is not present in "tree" format.
Similarly this information is included in old json
format for explain but not the new one.
There are 4 semijoin strategies used in MySQL,
FirstMatch, LooseScan, Materialization (Materialized
Lookup, Materialized Scan) and Duplicate weedout.
As of now, "Weedout" and "Materialization" (displayed
as Materialization with de-duplication) is already
displayed in the tree format. We add information about
"FirstMatch" and "LooseScan".
There is no difference between old optimizer and
hypergraph w.r.t "FirstMatch" strategy i.e all the
different ways of using the strategy, "Nested loop
semijoin", "Hash semijoin", "Inner join with limiting
"1" row from the inner side table" are all supported
even in hypergraph optimizer. However, in case of
"LooseScan", old optimizer always does it by using
an index. It also supports with multiple tables on
the inner side of the semijoin. In hypergraph,
the later is not supported. The former is supported
by using the interesting orders framework by
registering a "grouping".
Duplicate Weedout and Materialization strategies are
both not supported by hypergraph.

Change-Id: I1f77c5abcf335a88fb61fd64e2239875c8b390c8