Bug#36217437 PCA sees: Error OS signal received
in TUX, in function removeTuxEntries the TuxMaintReq req
is expected to have errorCode = 0 (no error).  but, sometimes,
errorCode in set by TUX (in execTUX_MAINT_REQ) causing data
node to crash due to an assertion failure on errorCode.

In execTUX_MAINT_REQ the number of different errorCode that
can be set is relatively big, so this patch adds an extra debug
with some useful information (inc. errorCode) to removeTuxEntries.

Similarly, extra debug is added to addTuxEntries as well.

Change-Id: Iceb61367e2869e466ece94c2c004a9ef24840f7e