Bug#36838373: Failure in Item::convert_charset()
The problem occurs when we perform type aggregation for a set operation,
which has input that is an Item_type_holder object from an underlying
set operation. Since Item_type_holder is never involved in regular
expression evaluation, it does not implement proper table dependency
information, thus used_tables() always return zero, and subsequently
it is regarded to be a const item. Since it is a string object and it
has collation different from the destination collation, a collation
conversion is attempted, which will call val_str(). But this function
is not implemented for the class, which is reasonable, given that the
class is not supposed to take part in evaluations. Hence the failure.

The fix is to return used_tables() that marks the object as non-const,
so we decide to return RAND_TABLE_BIT.

We also notice that the helper class Item_metadata_copy is redundant
and can be deleted. It is only used in type aggregation, but
Item_type_holder is sufficient for this. We must only ensure that type
aggregation is not performed into an Item_type_holder object until after
the type is completely determined, since Item_aggregate_type::join_types
may be called with itself as one of the arguments.
The Item_metadata_copy is removed and relevant functionality copied to
class Item_type_holder.

We also remove code and comments regarding rollback in
Item_aggregate_type::join_types(), since all transformations performed
here are permanent since we implemented prepare once.

Further on, the join_types() function is renamed as unify_types(), since
"join" is kind of a different operation.

Change-Id: Iae52edfe7fd7e0e0477f53190e5c962e0cc22f66