WL#16007: Subquery optimizer hints for hypergraph optimizer [2/3 SEMIJOIN]
Hypergraph optimizer does not support all the semijoin
strategies supported by the old optimizer. It supports
two strategies, FIRSTMATCH and LOOSESCAN. If both
are enabled for the query block, hypergraph will propose
both the plans if possible and the best choice will
be picked. If any one of them is enabled and if it
is possible to use the enabled semijoin strategy,
hypergraph will only propose that plan. First match
strategy is possible for all cases of semijoin however
loose scan is not. If loose scan is enabled and
first match is not, and if loose scan is not possible
the default semijoin strategy would be first match.

Materialization as a strategy for semijoins and also
duplicate weedout is not supported in hypergraph
optimizer. So any hints or switches, both enabling
or disabling w.r.t these two strategies will not
have the same effect as it will have on the old
optimizer. If these two startegies are disabled
using optimizer switches or through NO_SEMIJOIN()
hint, it will not have any effect on the hypergraph
optimizer. If they are used as part of the SEMIJOIN()
hint, if FIRSTMATCH or LOOSESCAN are not part
of the allowed startegies then hypergraph will by
default choose FIRSTMATCH as the strategy for this
case.

Change-Id: Ie72baa682e4ead79d273efd0e4cbd6a07bcb3c02