Bug#37039409 Backport Bug#36210202 FTS_DOC_ID comparison incorrect for large values
Problem:
The method of comparing doc ids in FTS is incorrect - to determine which
id is greater, first 64-bit unsigned subtraction is performed on them,
and then the result is cast to 32-bit signed integer. This leads to
incorrect results when the subtraction result is not in range of 32-bit
signed values. This leads to further incorrect behavior in FTS including
ranking score calculation.

Fix:
Subtraction "hack" replaced with simple comparison operators.

Patch based on contribution from Alibaba's Shaohua Wang. Thank you!

Change-Id: I5149d4de5de3c7e6732c1905fe54437ef7eb22e7