Bug#33334911 - Multi-valued index performance is too slow (Access too many rows than required)
Issue:
The query which has "greater or less than (> or <)" condition on
multi-valued index is too slow.

Analysis:
In case of query which has range conditions on the multi-valued
index, the storage engine kept reading the rows because key
comparison to check the end of the range was not getting executed.
Because of this it was reading more rows than expected leading to
slow execution of the query.

Fix:
Enable key comparison to check the end of the range in case of
multi-valued index. Disable multi-valued index for DS-MRR, except
for equality ranges, as virtual fields are not updated during the
index read in case of DS-MRR to do the key comparison.

Change-Id: I1e3019ab15b29615815c9279c85a3d7425701db0