WL#16482: mysqldump blindly trusting SHOW CREATE TABLE leads   to arbitrary code execution.
The 8.0 version.

Introduced a --[skip-]system-command command line/config file option
for the mysql binary.
Type: boolean
Default: ON, can be turned off with --skip-system-command

When --skip-system-command is specified and a system command or the
abbreviated \! is specified, an error will be printed and the command
won't be executed.

Test case added.

Change-Id: I46e8f3cc8e8a6aa37f8374c04a87177604cd124e