Bug#36739383 Assertion 'false' materialize_iterator::SpillState::read_next_row_secondary_over - part 4/4
This patch removes the old usage of overflow_memroot. With the new
space allocation scheme, cf.  reset_mem_root, we established that is
is safe to do so, since we no longer have fragmentation after round
one of using the hash table for set operations; the necessary space is
allocated as one big chunk. See Gerrit for more details.

Change-Id: Ifc125fa9789492696dad3c0866036d35cdfd4afe