Bug#36878077 Remove utf8/utf8mb3 alias handling from the strings library
utf8 is currently an alias for utf8mb3. We want to change this some
time in the future.

Remove the handling of alias names from the strings library, and
handle it in mysys instead. The idea is that MySQL code using mysys
will tell it how to interpret "utf8".

Also remove unused functions mysql::collation::find_default_binary().

Change-Id: I398ddc349e3f454468e5feeb14f66d818b09f75e