Bug#36886242 Node failure late in Index Statistics Update schema  transaction causes president failure
Problem

The Index Statistics UPDATE_STAT Schema Transaction can be
generated automatically by Ndb data nodes in response to
monitoring data change rates, or triggered as a part of
schema changes or manually by ANALYZE_TABLE et al.

Most Schema Transactions (ST) involve all data nodes equally
and so the survivable loss of one data node does not affect
the outcome, and the surviving nodes carry on.

The UPDATE_STAT ST has phases operating only on one node,
so that failure of that node does affect the overall outcome.

This is handled by requiring that any node failure during
the UPDATE_STAT ST results in the transaction being rolled
back.

However this mechanism did not handle the case where the
UPDATE_STAT ST was already in the short final commit or
complete phases.

In that case it caused an assertion failure at the
Master/President node.

That behaves as a cascading node failure - one node
failure leading to another, which endangers availability,
especially if the nodes involved are in the same nodegroup.

Solution

The UPDATE_STAT ST rollback on node failure behaviour is
limited to the pre-commit ST states.

This avoids the assertion failure and cascading node
failures.

Testing

A new testcase is added :

  testDict -n IndexStatNodeFailures T1

Change-Id: I5915087fa5d24b50e11bb5697e93400443a367d3