Bug#36882503 - Unexpected error response for mrs notes sample app authenticated request
Description
===========
As an MRS user, I expect to be able to create a new note using the MRS Notes sample app
as long as I'm able to sign in with an existing user for an MRS-based Authentication App.
This was the case for version 8.2.0-labs-mrs-7 of the MySQL Router, but it is no longer
the case for version 9.0.0-labs-mrs-8, where the Router returns back a 500 error response
for one of the requests sent by the client:

  Field 'user_id' doesn't have a default value

Fix
===
The `note` app configures MRS to translate column name from "user_id" to
"userId", where MRS used wrong name when searching for the column information
which leads to default handling of that column.

Change-Id: I1ac3ac1a709c877744b663e003c1942a63b4ecff