WL#16007: Subquery optimizer hints for hypergraph optimizer [3/3 SUBQUERY]
Support INTOEXISTS and MATERIALIZATION hint for subqueries.

Hypergraph optimizer has special processing for subqueries
which are transformed using "IN-TO-EXISTS" transformation.
It first proposes a plan for the subquery with conditions
added by the "IN-TO-EXISTS" transformation and then without
these conditions. It uses the second plan when it decides
to choose materialization for the subquery. Picking the
final strategy to be used happens at the end when the
plan is finalized (FinalizeMaterializedSubqueries()).
While deciding to materialize the subquery predicates attached
to a "FILTER" access path, hypergraph checks if the access
path is marked as having a materializable predicate. If there
is such a predicate, it checks if it could be materialized.
If so, it goes ahead and creates a plan with the subquery
not having the "IN-TO-EXISTS" conditions.
However, if the hint for the subquery forces materialization,
even if the "FILTER" access path is marked as not having
a materializable predicate, it needs to check if there
are subqueries in the condition attached to the "FILTER"
access path which force materialization. If so, it will
goahead and materialize the subquery if possible.
This change handles the "MATERIALIZATION" hint. For the
"INTOEXISTS" hint, a check is now added in the same
function, to see if the startegy is enforced. If so, it
does not materialize the subquery.
For a subquery present in projection, hypergraph does
not yet support materializing it. So hints will not have
the same behavior as that of old optimizer.

Change-Id: Ife6caa82df1f3889b044b32bd554b78d679e344f