Bug#31763497 PERFORMANCE DEGRADATION CAUSED BY MONITORING   SYS.INNODB_LOCK_WAITS IN MYSQL 8.0
PATCH for MySQL 8.4

Problem
=======

On a system in production, under load,
executing the monitoring query:

  SELECT * FROM sys.innodb_lock_waits;

cause performance degradation on the entire system,
affecting the server workload.

Root cause
==========

sys.innodb_lock_waits is a view,
defined as:
  SELECT ...
  FROM performance_schema.data_lock_waits w
  INNER JOIN information_schema.innodb_trx b ...
  INNER JOIN information_schema.innodb_trx r ...
  INNER JOIN performance_schema.data_locks bl ...
  INNER JOIN performance_schema.data_locks rl ...

For each wait (w) found in table performance_schema.data_lock_waits,
the view needs to lookup the corresponding (bl) and (rl) locks,
in table performance_schema.data_locks.

Issue 1

The view definition was not using the full primary key
when joining (w) with (bl), and (w) with (rl),
leading to a poor plan from the optimizer.

This issue has been addressed already with:
  Bug 36337708 View sys.innodb_lock_waits should join on primary keys

Issue 2

Once the optimizer produce a correct plan,
the primary key for table performance_schema.data_locks
is used to fetch a record directly, in a point select.

The implementation of table performance_schema.data_locks however,
was not performing a point select but a full scan instead,
in the underlying iterator used to inspect innodb lock data.

The scan not only make the query not efficient,
but as a side effect generates a lot of locks in innodb,
which slow down queries from the server workload.

Fix
===

Implement a proper PRIMARY KEY for table performance_schema.data_locks.

In particular, in:
  table_data_locks::index_next()
perform a primary key lookup and a point select,
instead of a table scan,
when the index is the PRIMARY KEY.

With this fix:
  SELECT * FROM sys.innodb_lock_waits
will only fetch 2 locks for each wait,
instead of scanning all locks twice for each wait.

Also, add the missing primary key in table performance_schema.data_lock_waits,
and implement it using the same pattern.

Change-Id: Ifd9294e985750515bf528d26d6dda6159a824157