Bug#36891161 Assignment from a local empty SQL_I_list got an illegal object
This patch is based on a contribution from Xingyu Yang <xingyuyang@tencent.com>

Problem:
========
SQL_I_list uses member field 'next' to track the pointer field to
the next object. It is initialized as the address of member field
'first', which is only valid with the list object constructed with
regular constructors.
The default assignment operator uses memberwise assignment, which
is not correct for an empty list.

Solution:
=========
Supply customized move constructor and operator= functions.

We also change the implementation of save_and_clear() so use std::move,
rather than assignment.

Change-Id: Ic47ab39e566d151a98dc178ea500130dbbd5c71b