Bug#36870544 Remove DEBUG_SYNC_C_IF_THD macro
Background
==========
DEBUG_SYNC_C_IF_THD macro was added to use current_thd in case of an
empty THD which was not the case with DEBUG_SYNC macro that time. Both
of these macros ends up calling the debug_sync() API.
At present DEBUG_SYNC_C_IF_THD macro is misleading, it makes callers
believe that it would use the THD passed in it, if the THD is empty
then, it would use the current_thd. But, if we look at the
implementation of callback method debug_sync_C_callback(), we notice
that it ignores the thd pointer but use the current_thd always.
This breaks the callers' assumptions that has caused, tests
instabilities and bugs.

This macro was added through commit#7b8b8dce in Feb 2012, but two
months later a null check was also added in the debug_sync() API
through commit#06aac18.  This means the following :
- There is no need to use DEBUG_SYNC_C_IF_THD macro, callers should be
  able express the intent clearly with DEBUG_SYNC macro.
- It is no-op if empty THD is passed to DEBUG_SYNC macro. We can detect
  the possible problem early and handle them in the callers.
- We could avoid bugs like Bug#108832 , Bug#113002.

Fix
===
Replaced the existing uses of DEBUG_SYNC_C_IF_THD macro with DEBUG_SYNC
and removed the macro.

Change-Id: Ib103c1d9aac035e56508f01e218375e8d71b66a4